refactor/#48: 기존 request 및 response network패키지에서 관리
