+++ app/src/main/java/com/example/yeongkkuel/network/request/login/ReissueRequest.kt
@@ -1,4 +1,4 @@
-package com.example.yeongkkuel.presentation.login.request
+package com.example.yeongkkuel.network.request.login
 
 data class ReissueRequest(
     val accessToken: String,
