+++ app/src/main/java/com/example/yeongkkuel/network/request/login/TermsAgreeRequest.kt
@@ -1,4 +1,4 @@
-package com.example.yeongkkuel.presentation.login.request
+package com.example.yeongkkuel.network.request.login
 
 data class TermsAgreeRequest(
     val term1: Boolean,
