+++ app/src/main/java/com/example/yeongkkuel/network/response/my/UserProfileResult.kt
@@ -1,4 +1,4 @@
-package com.example.yeongkkuel.presentation.my.data
+package com.example.yeongkkuel.network.response.my
 
 data class UserProfileResult(
     val nickname: String,
