+++ app/src/main/java/com/example/yeongkkuel/presentation/signup/SignupFragment.kt
@@ -19,8 +19,8 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.yeongkkuel.R
 import com.example.yeongkkuel.databinding.FragmentSignupBinding // 자동 생성된 바인딩 클래스
-import com.example.yeongkkuel.presentation.login.request.ReferralRequest
-import com.example.yeongkkuel.presentation.login.request.UserInfoRequest
+import com.example.yeongkkuel.network.request.login.ReferralRequest
+import com.example.yeongkkuel.network.request.login.UserInfoRequest
 import com.example.yeongkkuel.network.RetrofitClient
 import kotlinx.coroutines.launch
 import timber.log.Timber
